trace: move trace code to its own file
