+++ rust/ares/src/main.rs
@@ -104,4 +104,3 @@ fn main() -> io::Result<()> {
     };
     Ok(())
 }
-
