+++ rust/ares/src/serf.rs
@@ -1,6 +1,6 @@
 use crate::hamt::Hamt;
 use crate::interpreter;
-use crate::interpreter::{inc, interpret, Error, TraceInfo};
+use crate::interpreter::{inc, interpret, Error};
 use crate::jets::cold::Cold;
 use crate::jets::hot::Hot;
 use crate::jets::list::util::{lent, zing};
@@ -12,17 +12,16 @@ use crate::newt::Newt;
 use crate::noun::{Cell, Noun, Slots, D, T};
 use crate::snapshot::double_jam::DoubleJam;
 use crate::snapshot::Snapshot;
+use crate::trace::{create_trace_file, write_metadata, TraceInfo};
 use ares_macros::tas;
-use json::object;
 use signal_hook;
 use signal_hook::consts::SIGINT;
-use std::fs::{create_dir_all, File};
-use std::io::{self, Write};
+use std::fs::create_dir_all;
+use std::io;
 use std::path::PathBuf;
 use std::result::Result;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
-use std::time::Instant;
 
 crate::gdb!();
 
@@ -202,94 +201,15 @@ pub fn serf() -> io::Result<()> {
             "flag bitmap is not integer",
         )))?;
 
-    let trace_info = if wag & FLAG_TRACE != 0 {
-        let mut trace_dir_path = pier_path.clone();
-        trace_dir_path.push(".urb");
-        trace_dir_path.push("put");
-        trace_dir_path.push("trace");
-        create_dir_all(&trace_dir_path)?;
-
-        let trace_path: PathBuf;
-        let mut trace_idx = 0u32;
-        loop {
-            let mut prospective_path = trace_dir_path.clone();
-            prospective_path.push(format!("{}.json", trace_idx));
-
-            if prospective_path.exists() {
-                trace_idx += 1;
-            } else {
-                trace_path = prospective_path.clone();
-                break;
-            }
-        }
-
-        let mut file = File::create(trace_path)?;
-        let process_start = Instant::now();
-        let pid = std::process::id();
-
-        // write metadata to trace file
-
-        file.write_all("[ ".as_bytes())?;
-
-        (object! {
-            name: "process_name",
-            ph: "M",
-            pid: pid,
-            args: object! { name: "urbit", },
-        })
-        .write(&mut file)?;
-        file.write_all(",\n".as_bytes())?;
-
-        (object! {
-            name: "thread_name",
-            ph: "M",
-            pid: pid,
-            tid: 1,
-            args: object!{ name: "Event Processing", },
-        })
-        .write(&mut file)?;
-        file.write_all(",\n".as_bytes())?;
-
-        (object! {
-            name: "thread_sort_index",
-            ph: "M",
-            pid: pid,
-            tid: 1,
-            args: object!{ sort_index: 1, },
-        })
-        .write(&mut file)?;
-        file.write_all(",\n".as_bytes())?;
-
-        (object! {
-            name: "thread_name",
-            ph: "M",
-            pid: pid,
-            tid: 2,
-            args: object!{ name: "Nock", },
-        })
-        .write(&mut file)?;
-        file.write_all(",\n".as_bytes())?;
-
-        (object! {
-            name: "thread_sort_index",
-            ph: "M",
-            pid: pid,
-            tid: 2,
-            args: object!{ sort_index: 2, },
-        })
-        .write(&mut file)?;
-        file.write_all(",\n".as_bytes())?;
-
-        file.sync_data()?;
-
-        Some(TraceInfo {
-            file,
-            pid,
-            process_start,
-        })
+    let mut trace_info = if wag & FLAG_TRACE != 0 {
+        create_trace_file(pier_path).ok()
     } else {
         None
     };
+
+    if let Some(ref mut info) = trace_info.as_mut() {
+        write_metadata(info);
+    }
 
     let mut context = Context::new(&snap_path, trace_info);
     context.ripe();
